refactor: добавить небольшие изменения

+++ api/src/main/java/com/example/api/dto/DirectorDto.java
@@ -4,6 +4,6 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 
 @Schema(description = "Информация о режиссёре")
-public record DirectorDto(@Schema(description = "Идентификатор режиссёра") Integer id,
+public record DirectorDto(@Schema(description = "Идентификатор режиссёра") Long id,
                           @Schema(description = "Имя режиссёра") @NotBlank String name) {
 }
