+++ api/src/main/java/com/example/api/dto/EventDto.java
@@ -1,12 +1,14 @@
 package com.example.api.dto;
 
+import com.example.api.dto.enums.EventOperation;
+import com.example.api.dto.enums.EventType;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 @Schema(description = "Информация о действии пользователя")
-public record EventDto(@Schema(description = "Идентификатор действия") Integer eventId,
+public record EventDto(@Schema(description = "Идентификатор действия") Long eventId,
                        @Schema(description = "Время") Long timestamp,
-                       @Schema(description = "Идентификатор пользователя") Integer userId,
+                       @Schema(description = "Идентификатор пользователя") Long userId,
                        @Schema(description = "Тип действия") EventType eventType,
                        @Schema(description = "Тип операции") EventOperation operation,
-                       @Schema(description = "Идентификатор сущности") Integer entityId) {
+                       @Schema(description = "Идентификатор сущности") Long entityId) {
 }
