+++ api/src/main/java/com/example/api/dto/ReviewDto.java
@@ -5,10 +5,10 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
 @Schema(description = "Информация о рецензии")
-public record ReviewDto(@Schema(description = "Идентификатор рецензии") Integer reviewId,
+public record ReviewDto(@Schema(description = "Идентификатор рецензии") Long reviewId,
                         @Schema(description = "Содержимое рецензии") @NotBlank String content,
                         @Schema(description = "Является ли рецензия положительной") @NotNull Boolean isPositive,
-                        @Schema(description = "Идентификатор пользователя") @NotNull Integer userId,
-                        @Schema(description = "Идентификатор фильма") @NotNull Integer filmId,
+                        @Schema(description = "Идентификатор пользователя") @NotNull Long userId,
+                        @Schema(description = "Идентификатор фильма") @NotNull Long filmId,
                         @Schema(description = "Количество пользователей, посчитавших рецензию полезной") Integer useful) {
 }
