+++ api/src/main/java/com/example/api/dto/UserDto.java
@@ -9,10 +9,10 @@ import java.time.LocalDate;
 import java.util.List;
 
 @Schema(description = "Информация о пользователе")
-public record UserDto(@Schema(description = "Идентификатор пользователя") Integer id,
+public record UserDto(@Schema(description = "Идентификатор пользователя") Long id,
                       @Schema(description = "Электронная почта пользователя") @Email String email,
                       @Schema(description = "Login пользователя") @NotBlank String login,
                       @Schema(description = "Имя пользователя") String name,
                       @Schema(description = "День рождения пользователя") @Past LocalDate birthday,
-                      @Schema(description = "Список друзей пользователя") List<Integer> friends) {
+                      @Schema(description = "Список друзей пользователя") List<Long> friends) {
 }
