+++ backend/src/main/java/com/example/backend/dao/mapper/ReviewRowMapper.java
@@ -10,6 +10,6 @@ public class ReviewRowMapper implements RowMapper<Review> {
 
     @Override
     public Review mapRow(ResultSet rs, int rowNum) throws SQLException {
-        return new Review(rs.getInt("id"), rs.getString("content"), rs.getBoolean("is_positive"), rs.getInt("user_id"),  rs.getInt("film_id"), rs.getInt("useful"));
+        return new Review(rs.getLong("id"), rs.getString("content"), rs.getBoolean("is_positive"), rs.getLong("user_id"),  rs.getLong("film_id"), rs.getInt("useful"));
     }
 }
