+++ backend/src/main/java/com/example/backend/entity/User.java
@@ -9,10 +9,10 @@ import java.util.List;
 @AllArgsConstructor
 @Data
 public class User {
-    private Integer id;
+    private Long id;
     private final String email;
     private final String login;
     private final String name;
     private final LocalDate birthday;
-    private final List<Integer> friends;
+    private final List<Long> friends;
 }
