+++ backend/src/main/java/com/example/backend/web/MpaController.java
@@ -1,7 +1,7 @@
 package com.example.backend.web;
 
 import com.example.api.MpaApi;
-import com.example.api.dto.Mpa;
+import com.example.api.dto.enums.Mpa;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
@@ -9,7 +9,7 @@ import java.util.List;
 @RestController
 public class MpaController implements MpaApi {
 
-    public Mpa findById(Integer id) {
+    public Mpa findById(Long id) {
         return Mpa.findById(id);
     }
